Fix regression failures: float64 FFT precision + revert CUDA VAD
Two root causes of hash mismatches identified and fixed:

1. scipy.fft float32 precision: numpy.fft internally upcasts float32 input
   to float64 (pocketfft behaviour); scipy.fft respects input dtype and stays
   in float32, producing sub-ULP differences that accumulate over long audio.
   Fix: cast frames to float64 before scipy.fft call — preserves multi-threaded
   speedup (workers=-1) while producing bit-identical output to numpy.fft.

2. CUDA VAD non-determinism: SileroVAD on CUDAExecutionProvider produces
   slightly different speech-boundary timestamps for long audio due to GPU
   float non-determinism. Different boundaries → different chunk splits →
   different transcription. Fix: revert VAD to CPUExecutionProvider (deterministic).

Regression test expanded to 20 cases: full audio + 9 evenly-spaced 60s clips
× 2 configs (float16 bs=16 beam=5 and int8 bs=32 beam=1).

+++ faster_whisper/feature_extractor.py
@@ -203,12 +203,15 @@ class FeatureExtractor:
                     "Cannot have onesided output if window or input is complex"
                 )
             if _scipy_fft is not None:
-                output = _scipy_fft.fft(frames, n=n_fft, axis=-1, norm=norm, workers=_FFT_WORKERS)
+                # Cast to float64 to match numpy.fft internal precision (pocketfft
+                # upcasts float32→float64 internally; scipy respects input dtype).
+                # This preserves multi-threaded speedup while keeping bit-identical output.
+                output = _scipy_fft.fft(frames.astype(np.float64), n=n_fft, axis=-1, norm=norm, workers=_FFT_WORKERS)
             else:
                 output = np.fft.fft(frames, n=n_fft, axis=-1, norm=norm)
         else:
             if _scipy_fft is not None:
-                output = _scipy_fft.rfft(frames, n=n_fft, axis=-1, norm=norm, workers=_FFT_WORKERS)
+                output = _scipy_fft.rfft(frames.astype(np.float64), n=n_fft, axis=-1, norm=norm, workers=_FFT_WORKERS)
             else:
                 output = np.fft.rfft(frames, n=n_fft, axis=-1, norm=norm)
